requirements.txt
asgiref==3.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.15
django-cors-headers==4.4.0
django-filter==24.3
django-modeltranslation==0.18.11
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
django-rosetta==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
phonenumbers==8.13.43
pillow==10.2.0
polib==1.2.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==72.1.0
simplejson==3.19.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.2
